~buggy features
Need to estabilsh multiple lanes within a queue.
Yeilds are being executed after return causing unexpected behaviour.
Also returning from a parent isn't  stopping children's execution properly

+++ code/lib/hexidecimal.hpp
@@ -11,7 +11,7 @@
 #include <string>
 
 namespace Hexidecimal {
-  long int convert(std::string input);
+  unsigned long long convert(std::string input);
 }
